Fix implicit cast compile errors in CHQTS nanoapps
Bug: 73294389
Test: make -C {nanoapp} google_x86_linux, also verify CHQTS pass

+++ apps/chqts/src/general_test/basic_sensor_test_base.h
@@ -102,7 +102,7 @@ class BasicSensorTestBase : public Test {
   bool mExternalSamplingStatusChange;
   bool mDoneWithPassiveConfigure;
   State mState;
-  uint16_t mInstanceId;
+  uint32_t mInstanceId;
   uint32_t mSensorHandle;
   uint64_t mPreTimestamp;
   uint64_t mFirstEventTimestamp;
